shape model
Delete a whole lot of extraneous code from the "stats for shape" section where the multivariate shape analysis appears. Add a single model for shape that I think is the one that we want to go with. Increase number of iterations in the permutation procedure to improve estimates of p values.