Add captureMethod to statistical models
For the multivariate shape analysis and each of the univariate analyses, added a version of the statistical model that includes captureMethod (angling, fyke net, or electrofishing) as a predictor. Retained the previous statistical models and renamed them to include "Reduced" in their names. Compared the new models to the reduced models. Also some minor tidying to reduce line breaks.