Changes to univariate figures
Because we added a term for captureMethod to the statistical models, needed to re-work the calculation of the fits. Now using predict.merMod() to calculate the fits for a FN-captured fish in each lake.

Also some tweaks to size of figures and fonts, etc. in response to journal formatting guidelines.